NX-34: retain non-null type-81 identities and nonzero sequences
Entity51Frame, Entity51Record, and ParasolidEntity51Record now carry
NonNullXmt and NonZeroU32. Delete the independent xmt/sequence admission
predicate from the framer. Delta census and native identity formatting
convert to integers only at their output boundaries. The existing wire
adapter rejects null identities and zero sequences and preserves the exact
numeric JSON representation.

Two native test fixtures used sequence zero even though the type-81 grammar
requires a nonzero sequence. Give those fixtures sequence one; neither test
asserts sequence semantics. Existing output and source-value assertions are
unchanged.

Validation: cargo check -q -p cadmpeg-codec-nx --tests passed. Extended the
exact entity wire test with field-specific null/zero rejection cases.
Hook-Bypass-Reason: User requires the full gate once at mission completion.

+++ crates/cadmpeg-codec-nx/src/deltas/tests.rs
@@ -53,7 +53,7 @@ fn deltas_walks_complete_status_prefixed_entity_51_records() {
     let residual = crate::deltas::semantic_residual(&stream);
     let retained = crate::parasolid::entity_51_records(&residual);
     assert_eq!(retained.len(), 1);
-    assert_eq!(retained[0].xmt, 10);
+    assert_eq!(u32::from(retained[0].xmt), 10);
     assert!(residual[..stream.len()].iter().all(|byte| *byte == 0xff));
 
     stream[entity_len - 1] = 1;
